Add a minor perf improvement
