+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/mutable/SortNode.java
@@ -6,7 +6,6 @@ import com.oracle.truffle.api.TruffleLanguage.ContextReference;
 import com.oracle.truffle.api.dsl.Cached;
 import com.oracle.truffle.api.dsl.CachedContext;
 import com.oracle.truffle.api.dsl.Specialization;
-import com.oracle.truffle.api.frame.MaterializedFrame;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.profiles.ConditionProfile;
@@ -14,20 +13,15 @@ import java.util.Arrays;
 import java.util.Comparator;
 import org.enso.interpreter.Language;
 import org.enso.interpreter.dsl.BuiltinMethod;
-import org.enso.interpreter.node.callable.InvokeCallableNode;
-import org.enso.interpreter.node.callable.InvokeCallableNode.ArgumentsExecutionMode;
-import org.enso.interpreter.node.callable.InvokeCallableNode.DefaultsExecutionMode;
 import org.enso.interpreter.runtime.Context;
-import org.enso.interpreter.runtime.callable.argument.CallArgumentInfo;
 import org.enso.interpreter.runtime.callable.atom.Atom;
 import org.enso.interpreter.runtime.callable.atom.AtomConstructor;
 import org.enso.interpreter.runtime.data.Array;
 import org.enso.interpreter.runtime.error.PanicException;
-import org.enso.interpreter.runtime.state.Stateful;
-import org.enso.interpreter.runtime.state.data.EmptyMap;
 
 @BuiltinMethod(type = "Array", name = "sort", description = "Sorts a mutable array in place.")
 public abstract class SortNode extends Node {
+  private @Child ComparatorNode comparatorNode = ComparatorNode.build();
   private final ConditionProfile consProfile = ConditionProfile.createCountingProfile();
 
   abstract Object execute(VirtualFrame frame, Object _this, Object comparator);
@@ -41,13 +35,8 @@ public abstract class SortNode extends Node {
       VirtualFrame frame,
       Array _this,
       Object comparator,
-      @Cached("buildInvokeNode()") InvokeCallableNode invokeNode,
-      @CachedContext(Language.class) ContextReference<Context> ctxRef,
-      @Cached("ctxRef.get().getBuiltins().ordering().newLess()") Atom less,
-      @Cached("ctxRef.get().getBuiltins().ordering().newEqual()") Atom equal,
-      @Cached("ctxRef.get().getBuiltins().ordering().newGreater()") Atom greater) {
-    Comparator<Object> compare =
-        buildComparator(invokeNode, frame.materialize(), comparator, less, equal, greater, ctxRef);
+      @CachedContext(Language.class) ContextReference<Context> ctxRef) {
+    Comparator<Object> compare = (l, r) -> comparatorNode.execute(frame, comparator, l, r);
     doSort(_this.getItems(), compare);
     return ctxRef.get().getBuiltins().nothing().newInstance();
   }
@@ -71,38 +60,5 @@ public abstract class SortNode extends Node {
   @TruffleBoundary
   void doSort(Object[] items, Comparator<Object> compare) {
     Arrays.sort(items, compare);
-  }
-
-  Comparator<Object> buildComparator(
-      InvokeCallableNode invokeNode,
-      MaterializedFrame frame,
-      Object comparator,
-      Atom less,
-      Atom equal,
-      Atom greater,
-      ContextReference<Context> ctxRef) {
-    return (l, r) -> {
-      Stateful result =
-          invokeNode.execute(comparator, frame, EmptyMap.create(), new Object[] {l, r});
-      Object atom = result.getValue();
-      if (atom == less) {
-        return -1;
-      } else if (atom == equal) {
-        return 0;
-      } else if (atom == greater) {
-        return 1;
-      } else {
-        CompilerDirectives.transferToInterpreter();
-        var ordering = ctxRef.get().getBuiltins().ordering().ordering();
-        throw new PanicException(
-            ctxRef.get().getBuiltins().error().makeTypeError(ordering, result.getValue()), this);
-      }
-    };
-  }
-
-  InvokeCallableNode buildInvokeNode() {
-    CallArgumentInfo[] callArguments = {new CallArgumentInfo(), new CallArgumentInfo()};
-    return InvokeCallableNode.build(
-        callArguments, DefaultsExecutionMode.EXECUTE, ArgumentsExecutionMode.PRE_EXECUTED);
   }
 }
